Add delete functionality to individual post page

+++ app/routes/post.js
@@ -4,10 +4,14 @@ export default Ember.Route.extend({
   model(params) {
     return this.store.findRecord('post', params.post_id);
   },
+
   actions: {
-    destroyPost(post) {
-      post.destroyRecord();
-      this.transitionTo("index");
+    delete(post) {
+      // debugger;
+      if(confirm('Are you sure you want to delete this post?')) {
+        post.destroyRecord();
+        this.transitionTo("index");
+      }
     }
   }
 });
